Bring back the library targets.
As Mike Frysinger <vapier@gentoo.org> points out, "If you try to build in
 a subdir and the archive doesn't exist, you'll get an error about make
having no idea how to make "pcnet/libpcnet.a" and you then have to do it
by hand. With these targets, make knows which directory to change into and
how to generate the .a so you dont have to figure it out."

However we still kill the win32 ones since those refer to the old Cygwin
based port which no longer exists. Also apply a patch from Mike fixing up
the targets for some driver libs.